requirements.txt
#MongoDB
pymongo==4.3.3

# Core FastAPI
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
pydantic>=2.0.0,<3.0.0

# Network Monitoring
pysnmp>=4.4.12,<5.0.0
pythonping>=1.1.4,<2.0.0
python-nmap>=0.7.1,<0.8.0
dnspython>=2.4.2,<3.0.0

# Utilities
python-multipart>=0.0.6,<1.0.0
python-dateutil>=2.8.2,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
httpx>=0.25.1,<1.0.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
pywin32>=306,<308; sys_platform == 'win32'

# Performance
uvloop>=0.17.0,<1.0.0; sys_platform != 'win32'
httptools>=0.5.0,<1.0.0

# Additional for dashboard
jinja2>=3.0.0,<4.0.0
aiofiles>=23.1.0,<24.0.0

# Windows Authentication
python-ntlm3>=1.0.2,<2.0.0
